+++ django_mapengine/__init__.py
@@ -1,3 +1,3 @@
 """Map Engine init, holds version"""
 
-__version__ = "1.7.0"
+__version__ = "1.8.0"