+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
-## [Unreleased]
+## [1.8.0] - 2024-05-23
 ### Added
 - docs for popups